Adds init code
